restaurant_reco_phase5/tests/test_api.py
import os
from fastapi.testclient import TestClient
import pandas as pd
import pytest

from restaurant_reco_phase5.api import app, get_data

# Create a small dummy parquet file for testing
DUMMY_DATA_PATH = "tests/test_data_api.parquet"

@pytest.fixture(scope="module", autouse=True)
def setup_dummy_data():
    df = pd.DataFrame([
        {
            "restaurant_id": "api1", "name": "API Test Res", "address": "456 Test Blvd",
            "location": "Banashankari", "cuisines": ["North Indian"],
            "rating": 4.8, "votes": 200, "cost_for_two": 500,
            "online_order": True, "book_table": True, "url": "http://api-test.com"
        }
    ])
    df.to_parquet(DUMMY_DATA_PATH)
    yield
    if os.path.exists(DUMMY_DATA_PATH):
        os.remove(DUMMY_DATA_PATH)

def test_health_check():
    client = TestClient(app)
    response = client.get("/health")
    assert response.status_code == 200
    assert response.json()["status"] == "healthy"

def test_recommendations_endpoint():
    # We override the data loader for the test if possible,
    # or just ensure the file exists at the expected relative path.
    # For simplicity, we'll just run against the app and expect it to load data.
    client = TestClient(app)
    
    # Payload
    payload = {
        "place": "Banashankari",
        "top_n": 1,
        "debug": True
    }
    
    response = client.post("/recommendations", json=payload)
    
    assert response.status_code == 200
    data = response.json()
    
    assert "recommendations" in data
    assert "X-Process-Time" in response.headers
    
    # If the dummy data was loaded, we should see it
    if len(data["recommendations"]) > 0:
        assert "API Test Res" in [r["name"] for r in data["recommendations"]]
