restaurant_reco_phase5/pyproject.toml
 [tool.pytest.ini_options]
 testpaths = ["tests"]
 addopts = "-q"
 
[tool.poetry.dependencies]
 python = "^3.10"
 pandas = "2.2.3"
 pyarrow = "16.1.0"
 pydantic = "2.9.2"
 groq = ">=0.11.0"
 python-dotenv = ">=1.0.0"
 sentence-transformers = ">=5.1.2"
 scikit-learn = ">=1.6.1"
 fastapi = ">=0.128.8"
 uvicorn = ">=0.39.0"
